Rename Timeline component to Steps, implement step calculation logic, and integrate with TimerPanel
