+++ src/App.tsx
@@ -1,14 +1,12 @@
 import React, { useState, useEffect } from 'react';
-import { Container, Typography, Snackbar } from '@mui/material';
+import { Container, Typography } from '@mui/material';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import useMediaQuery from '@mui/material/useMediaQuery';
 import Header from './components/Header';
 import Settings from './components/Settings';
-import Controls from './components/Controls';
-import Timeline from './components/Steps';
+import Timeline from './components/Timeline';
 import Footer from './components/Footer';
 import { translations } from './translations/index'
-import { NotificationMode, Step } from './types';
 import './App.css';
 import {
   BrowserRouter,
@@ -19,8 +17,6 @@ import {
   Navigate,
   useParams
 } from 'react-router-dom';
-import { useTimer } from './hooks/useTimer';
-import { useWakeLock } from './hooks/useWakeLock';
 
 // Create theme with both light and dark modes
 const getTheme = (mode: 'light' | 'dark') => createTheme({
@@ -42,79 +38,6 @@ const getTheme = (mode: 'light' | 'dark') => createTheme({
   }
 });
 
-// Function to calculate timer steps based on the new hybrid method
-function calculateSteps(beansAmount: number, flavor: string) {
-  // Total water used = beansAmount * 15
-  const totalWater = beansAmount * 15;
-  const flavorWater = totalWater * 0.4;
-  const strengthWater = totalWater * 0.6;
-  let flavor1, flavor2;
-  // Adjust flavor pours based on taste selection
-  if (flavor === "sweet") {
-    flavor1 = flavorWater * 0.42;
-    flavor2 = flavorWater * 0.58;
-  } else if (flavor === "sour") {
-    flavor1 = flavorWater * 0.58;
-    flavor2 = flavorWater * 0.42;
-  } else {
-    flavor1 = flavorWater * 0.5;
-    flavor2 = flavorWater * 0.5;
-  }
-  // Fixed strength steps
-  let strengthSteps = 2;
-
-  const steps: Array<Step> = [];
-  // Flavor pours are fixed at 0s and 45s
-  steps.push({
-    time: 0,
-    pourAmount: flavor1,
-    cumulative: flavor1,
-    descriptionKey: "flavorPour1",
-    status: 'upcoming'
-  });
-  steps.push({
-    time: 40,
-    pourAmount: flavor2,
-    cumulative: flavor1 + flavor2,
-    descriptionKey: "flavorPour2",
-    status: 'upcoming'
-  });
-  // Strength pour 1 is fixed at 90 seconds (1:30)
-  const strengthPourAmount = strengthWater / strengthSteps;
-  steps.push({
-    time: 90,
-    pourAmount: strengthPourAmount,
-    cumulative: steps[steps.length - 1].cumulative + strengthWater * 0.444,
-    descriptionKey: "strengthPour1",
-    status: 'upcoming'
-  });
-  // Strength pour 2 is fixed at 130 seconds (2:10)
-  steps.push({
-    time: 130,
-    pourAmount: strengthPourAmount,
-    cumulative: steps[steps.length - 1].cumulative + strengthWater * 0.556,
-    descriptionKey: "strengthPour2",
-    status: 'upcoming'
-  });
-  // Final step (finish) is fixed at 210 seconds
-  steps.push({
-    time: 165,
-    pourAmount: 0,
-    cumulative: totalWater,
-    descriptionKey: "open",
-    status: 'upcoming'
-  });
-  // Final step (finish) is fixed at 210 seconds
-  steps.push({
-    time: 210,
-    pourAmount: 0,
-    cumulative: totalWater,
-    descriptionKey: "finish",
-    status: 'upcoming'
-  });
-  return steps;
-}
-
 const getUserLang = () => {
   const userLang = navigator.language || navigator.languages[0];
   return userLang.startsWith('ja') ? 'ja' : 'en';
@@ -134,8 +57,6 @@ function AppWrapper() {
 }
 
 function App() {
-  const { currentTime, isRunning, start, pause, reset } = useTimer();
-  const { requestWakeLock, releaseWakeLock } = useWakeLock();
   const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)');
   const [darkMode, setDarkMode] = useState(prefersDarkMode);
   const theme = getTheme(darkMode ? 'dark' : 'light');
@@ -143,20 +64,24 @@ function App() {
   const t = translations[language]; // shorthand for current translations
   const [beansAmount, setBeansAmount] = useState(20);
   const [flavor, setFlavor] = useState("middle");
-  const [steps, setSteps] = useState<Step[]>([]);
   const [searchParams, setSearchParams] = useSearchParams();
-  const [notificationMode, setNotificationMode] = useState<NotificationMode>('none');
-  const [voice, setVoice] = useState<'male' | 'female'>('female');
   const navigate = useNavigate();
   const { lang } = useParams();
-  const [snackbarOpen, setSnackbarOpen] = useState(false);
-  
+
   useEffect(() => {
     if (lang && (lang === 'en' || lang === 'ja')) {
       setLanguage(lang);
     }
   }, [lang]);
 
+  // Handler for language toggle
+  const handleLanguageChange = (_e: React.MouseEvent<HTMLElement>, newLang: "en" | "ja") => {
+    if (newLang) {
+      setLanguage(newLang);
+      navigate(`/${newLang}/recipes/new-hybrid-method`);
+    }
+  };
+  
   useEffect(() => {
     const paramBeans = searchParams.get('beans');
     const paramFlavor = searchParams.get('flavor');
@@ -184,48 +109,12 @@ function App() {
       params.set('flavor', flavor);
       setSearchParams(params, { replace: true });
     }
-
-    const newSteps = calculateSteps(beansAmount, flavor);
-    setSteps(newSteps);
   }, [beansAmount, flavor]);
 
   // Update dark mode when system preference changes
   useEffect(() => {
     setDarkMode(prefersDarkMode);
   }, [prefersDarkMode]);
-
-  const handlePlay = async () => {
-    if (isRunning) return;
-    setSnackbarOpen(true);
-    await requestWakeLock();
-    start();
-  };
-
-  const handlePause = async () => {
-    setSnackbarOpen(false);
-    await releaseWakeLock();
-    pause();
-  };
-
-  // Reset the timer
-  const handleReset = async () => {
-    setSnackbarOpen(false);
-    await releaseWakeLock();
-    reset();
-  };
-
-  // Handler for language toggle
-  const handleLanguageChange = (_e: React.MouseEvent<HTMLElement>, newLang: "en" | "ja") => {
-    if (newLang) {
-      setLanguage(newLang);
-      navigate(`/${newLang}/recipes/new-hybrid-method`);
-    }
-  };
-
-  // Snackbarを閉じるハンドラー
-  const handleSnackbarClose = () => {
-    setSnackbarOpen(false);
-  };
 
   return (
     <ThemeProvider theme={theme}>
@@ -260,6 +149,7 @@ function App() {
             </a>
           )}
         </Typography>
+        
         <Settings
           t={t}
           beansAmount={beansAmount}
@@ -286,37 +176,15 @@ function App() {
           ))}
         </Typography>
 
-        <Controls
-          t={t}
-          onPlay={handlePlay}
-          onPause={handlePause}
-          onReset={handleReset}
-          notificationMode={notificationMode}
-          setNotificationMode={setNotificationMode}
-          voice={voice}
-          setVoice={setVoice}
-        />
-
         <Timeline
           t={t}
-          steps={steps}
-          setSteps={setSteps}
-          currentTime={currentTime}
           darkMode={darkMode}
-          notificationMode={notificationMode}
           language={language}
-          voice={voice}
-          onTimerComplete={handlePause}
+          beansAmount={beansAmount}
+          flavor={flavor}
         />
 
         <Footer t={t} />
-        <Snackbar
-          open={snackbarOpen}
-          autoHideDuration={5000}
-          onClose={handleSnackbarClose}
-          message={t.keepScreenOn}
-          anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
-        />
       </Container>
     </ThemeProvider>
   );
